Attempt to install during runtime
